Revert "adding new classes, folders"
