+++ Lab21.sln
@@ -3,9 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.33516.290
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Lab21", "prim1\Lab21.csproj", "{A706882D-3B2A-4882-BB70-57C8023CC877}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SweetGift", "SweetGift", "{02969ABD-F0EB-48ED-9BFF-EAD2114EFBDF}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lab21", "prim1\Lab21.csproj", "{A706882D-3B2A-4882-BB70-57C8023CC877}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -20,9 +18,6 @@ Global
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{A706882D-3B2A-4882-BB70-57C8023CC877} = {02969ABD-F0EB-48ED-9BFF-EAD2114EFBDF}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {D43A376A-C0A1-44DF-89EB-186755470D75}
