+++ prim1/View.cs
@@ -1,14 +1,34 @@
-﻿// buisness logics
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Lab21.Model
+namespace Lab21
 {
-    internal class Manager
+    public class Gift
     {
+        public static void PrintGiftList(Sweet[] sweets)
+        {
+            Console.WriteLine("Gift List:");
+            Console.WriteLine("-----------");
+
+            for (int i = 0; i < 9; i++)
+            {
+                Sweet sweet = sweets[i];
+                Console.Write("{0}: {1} {2} byn", i + 1, sweet.Name, sweet.Price);
+                if (sweet is Cookie cookie)
+                {
+                    Console.WriteLine(" {0}", cookie.IsChocolate ? "chocolate" : "non-chocolate");
+                }
+                else if (sweet is Candy candy)
+                {
+                    Console.WriteLine(" {0}", candy.IsSugarFree ? "sugar-free" : "full of sugar");
+                }
+            }
+
+        }
+
         public static Sweet FindCheapestSweet(Sweet[] sweets)
         {
             Sweet cheapestSweet = sweets[0];
